requirements.txt
streamlit==1.29.0
tensorflow==2.15.0
numpy==1.26.2
opencv-python==4.8.1
Pillow==10.0.1
albumentations==1.3.1